+++ cpp/src/groupby/groupby.cu
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-#include <reductions/histogram_helpers.hpp>
-
 #include <cudf/column/column.hpp>
 #include <cudf/column/column_factories.hpp>
 #include <cudf/column/column_view.hpp>
@@ -26,6 +24,7 @@
 #include <cudf/detail/groupby.hpp>
 #include <cudf/detail/groupby/group_replace_nulls.hpp>
 #include <cudf/detail/groupby/sort_helper.hpp>
+#include <cudf/detail/histogram_helpers.hpp>
 #include <cudf/detail/nvtx/ranges.hpp>
 #include <cudf/detail/utilities/vector_factories.hpp>
 #include <cudf/dictionary/dictionary_column_view.hpp>
